add algorithm base
Former-commit-id: 804068341f822cb65ec087871cb9604ecb9ffa50

+++ rover/src/rover/applications/aid/VruAid.cc
@@ -54,7 +54,8 @@ Coord VruAid::getCurrentLocation() {
 
 void VruAid::setAppRequirements() {
   L3Address destAddr = chooseDestAddr();
-  socket.setAppRequirements(par("minRate"), par("maxRate"),
+  socket.setAppRequirements(par("minRate").doubleValue(),
+                            par("maxRate").doubleValue(),
                             AidRecipientClass::ALL_LOCAL, destAddr, destPort);
 }
